C Functions, nested loops

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,16 +8,16 @@
 */
 int main(void)
 {
-int fib1 = 1, fib2 = 2, nextTerm;
+unsigned long long fib1 = 1, fib2 = 2, nextTerm;
 int count;
 
-printf("%d, %d", fib1, fib2);
+printf("%llu, %llu", fib1, fib2);
 
 for (count = 3; count <= 98; ++count)
 {
 nextTerm = fib1 + fib2;
 
-printf(", %d", nextTerm);
+printf(", %llu", nextTerm);
 
 fib1 = fib2;
 fib2 = nextTerm;
